requirements.txt
pandas
requests
python-dotenv
mysql-connector-python
pathlib
